backend/requirements.txt
# PostgreSQL database adapter
psycopg2-binary>=2.9.9

# Google Generative AI for embeddings
google-generativeai>=0.8.0

# Environment variables
python-dotenv>=1.0.0

# Optional: for better performance
numpy>=1.24.0
